requirements.txt
alabaster==0.7.12
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.8.2
bs4==0.0.1
cenpy==1.0.0.post2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
coverage==5.0.3
cycler==0.10.0
docutils==0.16
entrypoints==0.3
feather-format==0.4.0
Fiona==1.8.13
flake8==3.7.9
fuzzywuzzy==0.18.0
geopandas==0.7.0
idna==2.8
imagesize==1.2.0
Jinja2==2.11.1
kiwisolver==1.1.0
libpysal==4.2.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
munch==2.5.0
numpy==1.18.1
packaging==20.1
pandas==1.0.1
pyarrow==0.16.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pyparsing==2.4.6
pyproj==2.4.2.post1
python-dateutil==2.8.1
python-dotenv==0.11.0
python-Levenshtein==0.12.0
pytz==2019.3
requests==2.22.0
Rtree==0.9.4
scipy==1.4.1
seaborn==0.10.0
Shapely==1.7.0
six==1.14.0
slugify==0.0.1
snowballstemmer==2.0.0
soupsieve==1.9.5
Sphinx==2.4.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
tqdm==4.42.1
urllib3==1.25.8
xlrd==1.2.0
